+++ changelog.md
@@ -9,6 +9,9 @@
 - Plugin Comments agora tem uma interface de moderação na admin
 
 ### Added
+- Descrição para todos os hooks do sistema adicionadas
+- Interface de hooks adicionado à admin
+- Plugin Banners adicionado
 - Agora getSettingsDefinitions, settingsAll e settingsGroup tem a opção de pegar apenas settings do sistema, evitando as injetadas por plugins e temas
 
 ## [1.0.0] 2026-07-12